requirements.txt
Django==5.1.3
django-simple-captcha==0.6.0
django-session-timeout==0.1.0
mysqlclient==2.2.6
Pillow==11.0.0
python-dateutil==2.9.0.post0
asgiref==3.8.1
django-ranged-response==0.2.0
tzdata==2024.2
xlwt==1.3.0